refactor(adapter): move Claude loader under adapters
Move the Claude usage loader into the adapter module tree so it follows the same runtime layout as the other agent implementations.

Keep the existing root re-exports for shared loader helpers, but point them at adapter::claude and update the all-agent report path to call the adapter module directly. This is a layout-only change; the loader contents are unchanged.
