+++ rust/crates/ccusage/src/adapter/mod.rs
@@ -1,5 +1,6 @@
 pub(crate) mod all;
 pub(crate) mod amp;
+pub(crate) mod claude;
 pub(crate) mod codebuff;
 pub(crate) mod codex;
 pub(crate) mod copilot;
